initDerivedValues
